blog/pyproject.toml
[project]
name = "blog"
version = "0.1.0"
description = ""
requires-python = ">=3.13"
dependencies = [
    "aiosmtplib>=4.0.0",
    "fastapi[standard]>=0.115.12",
    "jinja2==3.1.5",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.4.0",
]
